refactor: move color helpers to separate file (#47)
